+++ source/classes/CosmosDB/CosmosDB.cs
@@ -1,7 +1,10 @@
 ﻿using System;
+using System.Net;
 
-namespace CosmosDB {
-    public enum Environment {
+namespace CosmosDB
+{
+    public enum Environment
+    {
         AzureChinaCloud,
         AzureCloud,
         AzureUSGovernment
@@ -32,25 +35,31 @@ namespace CosmosDB {
         public CosmosDB.ContextToken[] Token { get; set; }
         public System.Security.SecureString EntraIdToken { get; set; }
         public CosmosDB.BackoffPolicy BackoffPolicy { get; set; }
-        public CosmosDB.Environment Environment  { get; set; } = Environment.AzureCloud;
+        public CosmosDB.Environment Environment { get; set; } = Environment.AzureCloud;
     }
 
-    namespace IndexingPolicy {
-        namespace Path {
-            public class Index {
+    namespace IndexingPolicy
+    {
+        namespace Path
+        {
+            public class Index
+            {
                 public System.String dataType { get; set; }
                 public System.String kind { get; set; }
             }
 
-            public class IndexRange : CosmosDB.IndexingPolicy.Path.Index {
+            public class IndexRange : CosmosDB.IndexingPolicy.Path.Index
+            {
                 public readonly System.Int32 precision = -1;
             }
 
-            public class IndexHash : CosmosDB.IndexingPolicy.Path.Index {
+            public class IndexHash : CosmosDB.IndexingPolicy.Path.Index
+            {
                 public readonly System.Int32 precision = -1;
             }
 
-            public class IndexSpatial : CosmosDB.IndexingPolicy.Path.Index {
+            public class IndexSpatial : CosmosDB.IndexingPolicy.Path.Index
+            {
             }
 
             public class IncludedPath
@@ -70,7 +79,8 @@ namespace CosmosDB {
         }
 
 
-        namespace CompositeIndex {
+        namespace CompositeIndex
+        {
             public class Element
             {
                 public System.String path { get; set; }
@@ -88,14 +98,35 @@ namespace CosmosDB {
         }
     }
 
-    namespace UniqueKeyPolicy {
-        public class UniqueKey {
+    namespace UniqueKeyPolicy
+    {
+        public class UniqueKey
+        {
             public System.String[] paths { get; set; }
         }
 
         public class Policy
         {
             public CosmosDB.UniqueKeyPolicy.UniqueKey[] uniqueKeys { get; set; }
+        }
+    }
+
+    // ResponseException is used to handle exceptions that occur during the invocation of CosmosDB operations.
+    // Todo: Change to be based on Microsoft.PowerShell.Commands.HttpResponseException in 7.0.0.
+    // This will prevent this module from working in Windows PowerShell 5.1 because this class doesn't exist.
+    // Proposal is to drop support for Windows PowerShell 5.1 in 7.0.0.
+    public class ResponseException : System.Exception
+    {
+        public ResponseException(System.String message) : base(message) { }
+        public ResponseException(System.String message, System.Net.HttpStatusCode statusCode) : base(message)
+        {
+            this.StatusCode = statusCode;
+        }
+        public System.Net.HttpStatusCode? StatusCode { get; set; } = null;
+        public System.String Response { get; set; } = "";
+        public override System.String ToString()
+        {
+            return this.Message;
         }
     }
 }